# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Elder {
  name: string;
  age: number;
  relationship: string;
  location: string;
  avatarUrl: string;
  healthStatus: string;
}

export interface DashboardSummary {
  elder: Elder;
  todayProgress: number;
  todayTotal: number;
  medicationProgress: number;
  medicationTotal: number;
  trustScore: number;
  activeAlerts: number;
  monthlySpend: number;
  nextCheckIn: string;
  nextCheckInTime: string;
}

export interface Activity {
  id: number;
  type: string;
  title: string;
  detail: string;
  time: string;
  status: string;
  initials: string;
}

export interface CareLog {
  id: number;
  title: string;
  detail: string;
  time: string;
  category: string;
  status: string;
  caregiver: string;
  /** @nullable */
  photoUrl: string | null;
  /** @nullable */
  verifiedBy: string | null;
}

export interface CareLogInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  detail: string;
  category: string;
  time: string;
}

export interface CareLogUpdate {
  status: string;
  verifiedBy?: string;
}

export interface Medication {
  id: number;
  name: string;
  dosage: string;
  time: string;
  instruction: string;
  status: string;
  stock: number;
  stockLabel: string;
}

export interface MedicationStatusInput {
  status: string;
}

export interface Caregiver {
  id: number;
  name: string;
  role: string;
  initials: string;
  trustScore: number;
  status: string;
  distance: string;
}

export interface ScheduleItem {
  id: number;
  time: string;
  title: string;
  detail: string;
  status: string;
}

export interface Expense {
  id: number;
  merchant: string;
  detail: string;
  amount: number;
  date: string;
  status: string;
  /** @nullable */
  receiptUrl: string | null;
}

export interface ExpenseInput {
  /** @minLength 1 */
  merchant: string;
  /** @minLength 1 */
  detail: string;
  /** @minimum 0 */
  amount: number;
  date: string;
}

export interface SosInput {
  source: string;
  note: string;
}

export interface SosAlert {
  id: number;
  createdAt: string;
  status: string;
  notified: string[];
}

export interface ProfileDraftInput {
  /** @minLength 1 */
  transcript: string;
}

export interface ProfileDraft {
  summary: string;
  experience: string;
  skills: string[];
  availability: string;
  location: string;
}