# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  type: string;
  title: string;
  detail: string;
  time: string;
  status: string;
  initials: string;
}