# lib/api-zod/src/generated/types/careLog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareLog {
  id: number;
  title: string;
  detail: string;
  time: string;
  category: string;
  status: string;
  caregiver: string;
  /** @nullable */
  photoUrl: string | null;
  /** @nullable */
  verifiedBy: string | null;
}