# lib/api-zod/src/generated/types/careLogInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareLogInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  detail: string;
  category: string;
  time: string;
}