# lib/api-zod/src/generated/types/caregiver.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Caregiver {
  id: number;
  name: string;
  role: string;
  initials: string;
  trustScore: number;
  status: string;
  distance: string;
}