# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Elder } from './elder';

export interface DashboardSummary {
  elder: Elder;
  todayProgress: number;
  todayTotal: number;
  medicationProgress: number;
  medicationTotal: number;
  trustScore: number;
  activeAlerts: number;
  monthlySpend: number;
  nextCheckIn: string;
  nextCheckInTime: string;
}