# lib/api-zod/src/generated/types/elder.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Elder {
  name: string;
  age: number;
  relationship: string;
  location: string;
  avatarUrl: string;
  healthStatus: string;
}