# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  merchant: string;
  detail: string;
  amount: number;
  date: string;
  status: string;
  /** @nullable */
  receiptUrl: string | null;
}