# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  /** @minLength 1 */
  merchant: string;
  /** @minLength 1 */
  detail: string;
  /** @minimum 0 */
  amount: number;
  date: string;
}