# lib/api-zod/src/generated/types/medication.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Medication {
  id: number;
  name: string;
  dosage: string;
  time: string;
  instruction: string;
  status: string;
  stock: number;
  stockLabel: string;
}