# lib/api-zod/src/generated/types/profileDraft.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileDraft {
  summary: string;
  experience: string;
  skills: string[];
  availability: string;
  location: string;
}