# lib/api-zod/src/generated/types/profileDraftInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileDraftInput {
  /** @minLength 1 */
  transcript: string;
}