# lib/api-zod/src/generated/types/scheduleItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleItem {
  id: number;
  time: string;
  title: string;
  detail: string;
  status: string;
}