# lib/api-zod/src/generated/types/sosAlert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SosAlert {
  id: number;
  createdAt: string;
  status: string;
  notified: string[];
}